requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.11.17
cffi==1.16.0
cryptography==41.0.5
Django==3.2.9
django-cors-headers==4.4.0
django-dotenv==1.4.2
django-nine==0.2.7
djangorestframework==3.14.0
mysqlclient==2.2.0
packaging==23.2
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.8.0
urllib3==2.1.0
